// src/dorc-web/src/apis/dorc-api/models/PropertyApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface PropertyApiModel
 */
export interface PropertyApiModel {
  /**
   * @type {number}
   * @memberof PropertyApiModel
   */
  Id?: number;
  /**
   * @type {string}
   * @memberof PropertyApiModel
   */
  Name?: string | null;
  /**
   * @type {boolean}
   * @memberof PropertyApiModel
   */
  Secure?: boolean;
  /**
   * @type {boolean}
   * @memberof PropertyApiModel
   */
  IsArray?: boolean;
}