// src/dorc-web/src/apis/dorc-api/models/RequestProperty.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface RequestProperty
 */
export interface RequestProperty {
  /**
   * @type {string}
   * @memberof RequestProperty
   */
  PropertyName?: string | null;
  /**
   * @type {string}
   * @memberof RequestProperty
   */
  PropertyValue?: string | null;
}